+++ docs/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Sync existing Jira trackers that were added before Bugzilla sync (OSIDB-4736)
 
 ## [5.4.1] - 2026-01-15
 ### Removed